proper data being displayed

+++ app/forms/activity_form.py
@@ -6,9 +6,9 @@ from wtforms.validators import DataRequired, ValidationError, NumberRange, Lengt
 class ActivityForm(FlaskForm):
     user_id = IntegerField("User id", [DataRequired("User Id required")])
     distance = FloatField("Distance", [Optional(), NumberRange(min=0)])
-    hours = FloatField("Hours", [Optional(), NumberRange(min=0)])
-    minutes = FloatField("Minutes", [Optional(), NumberRange(min=0, max=60)])
-    seconds = FloatField("Seconds", [Optional(), NumberRange(min=0, max=60)])
+    hours = IntegerField("Hours", [Optional(), NumberRange(min=0)])
+    minutes = IntegerField("Minutes", [Optional(), NumberRange(min=0, max=60)])
+    seconds = IntegerField("Seconds", [Optional(), NumberRange(min=0, max=60)])
     elevation = FloatField("Elevation", [Optional(), NumberRange(max=29029)])
     extertion = IntegerField(
         "Extertion level", [Optional(), NumberRange(min=1, max=10)])
